examples/langgraph-viral-youtube/requirements.txt
langgraph>=0.2.0
langchain-anthropic>=0.3.0
langchain-core>=0.3.0
tavily-python>=0.5.0
bmasterai>=0.2.3
python-dotenv>=1.0.0
